--- backend/module-common/src/main/java/com/toonpick/exception/JwtException.java
@@ -1,10 +1,0 @@
-package com.toonpick.exception;
-
-/**
- * JWT 토큰 예외
- */
-public class JwtException extends RuntimeException {
-    public JwtException(String message) {
-        super(message);
-    }
-}
