+++ backend/module-internal/core-security/src/main/java/com/toonpick/exception/ExpiredJwtTokenException.java
@@ -1,0 +1,29 @@
+package com.toonpick.exception;
+
+import com.toonpick.type.ErrorCode;
+
+/**
+ * 만료되어 유효하지 않는 JWT Token 예외
+ */
+public class ExpiredJwtTokenException extends JwtException {
+
+    public ExpiredJwtTokenException(String message) {
+        super(message);
+    }
+
+    public ExpiredJwtTokenException(ErrorCode errorCode) {
+        super(errorCode);
+    }
+
+    public ExpiredJwtTokenException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public ExpiredJwtTokenException(ErrorCode errorCode, Throwable cause) {
+        super(errorCode, cause);
+    }
+
+    public ExpiredJwtTokenException(ErrorCode errorCode, String message) {
+        super(errorCode, message);
+    }
+}
