+++ backend/module-internal/core-security/src/main/java/com/toonpick/exception/MissingJwtTokenException.java
@@ -1,0 +1,29 @@
+package com.toonpick.exception;
+
+import com.toonpick.type.ErrorCode;
+
+/**
+ * 요청에서 Jwt 토큰을 찾을 수 없는 경우 예외
+ */
+public class MissingJwtTokenException extends JwtException {
+
+    public MissingJwtTokenException(String message) {
+        super(message);
+    }
+
+    public MissingJwtTokenException(ErrorCode errorCode) {
+        super(errorCode);
+    }
+
+    public MissingJwtTokenException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public MissingJwtTokenException(ErrorCode errorCode, Throwable cause) {
+        super(errorCode, cause);
+    }
+
+    public MissingJwtTokenException(ErrorCode errorCode, String message) {
+        super(errorCode, message);
+    }
+}
